* OK, I could have *sworn* that ecore was supposed to delete your client socket automatically on disconnect - apparantly not.  For those having issues with epsilon crashing after 'a while' - this fixes it.  Process was hitting FD limit
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@31055 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ e17/libs/epsilon/src/lib/epsilon_thumb.c
@@ -14,7 +14,7 @@ EAPI int EPSILON_EVENT_DONE;
 static int debug = 0;
 
 static int epsilon_init_count = 0;
-static int epsilon_mid = 0;
+static unsigned long epsilon_mid = 0;
 
 static Ecore_DList *epsilon_request_queue = NULL;
